pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama-vision-mcp"
version = "1.0.0"
description = "A Model Context Protocol server providing computer vision capabilities using Ollama"
authors = [
    {name = "Ollama Vision MCP Contributors"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
keywords = ["mcp", "ollama", "vision", "computer-vision", "ai", "llm"]

dependencies = [
    "mcp>=0.9.1",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[project.urls]
"Homepage" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp"
"Bug Reports" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp/issues"
"Source" = "https://github.com/ollama-vision-mcp/ollama-vision-mcp"

[project.scripts]
ollama-vision-mcp = "src.server:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
